Created separate classes for each function
